Pulled SampleDistribution method out to shared utils
